+++ src/auth/authFiles/auth.ts
@@ -1,0 +1,15 @@
+import { WorkspaceConfiguration } from "vscode";
+
+export const authRouteFileContent = (config: WorkspaceConfiguration) => {
+  const runtimeDependency = config.get("runtimeDependency") || "@remix-run/node";
+  return [
+    `import type { LoaderArgs } from "${runtimeDependency}";`,
+    'import { authenticator } from "~/services/auth.server";',
+    "",
+    "export const loader = async ({ request }: LoaderArgs) => {",
+    "  return await authenticator.isAuthenticated(request, {",
+    '    successRedirect: "/dashboard",',
+    "  });",
+    "};",
+  ].join("\n");
+};